requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
sqlalchemy>=1.4.23
aiohttp>=3.8.1
python-telegram-bot>=13.7
qbittorrent-api>=2023.3.44
PyYAML>=6.0
python-multipart>=0.0.5
requests>=2.26.0
tmdbv3api>=1.7.7
beautifulsoup4>=4.9.3
lxml>=4.9.0
aiosqlite>=0.17.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bencodepy
apscheduler